+++ src/main/java/io/github/mivek/internationalization/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * @author mivek
+ *
+ */
+package io.github.mivek.internationalization;
