--- src/main/java/io/github/mivekinternationalization/package-info.java
@@ -1,5 +1,0 @@
-/**
- * @author mivek
- *
- */
-package io.github.mivekinternationalization;
