+++ src/test/java/io/github/mivek/facade/MetarFacadeTest.java
@@ -11,8 +11,8 @@ import static org.junit.Assert.assertTrue;
 import org.junit.Test;
 
 import io.github.mivek.exception.ParseException;
+import io.github.mivek.internationalization.Messages;
 import io.github.mivek.model.Metar;
-import io.github.mivekinternationalization.Messages;
 
 public class MetarFacadeTest extends AbstractWeatherCodeFacadeTest<Metar> {
     @Test
